Code updated to separate dao and service
