+++ src/main/java/com/app/spb/constants/SpbConstants.java
@@ -1,0 +1,9 @@
+package com.app.spb.constants;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public class SpbConstants {
+	public static final String SUCCESS = "SUCCESS";
+}
