+++ src/main/java/com/app/spb/controller/ProductController.java
@@ -11,8 +11,8 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.app.spb.dao.SpbDao;
 import com.app.spb.entity.Product;
-import com.app.spb.service.ProductService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import lombok.RequiredArgsConstructor;
@@ -21,7 +21,7 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class ProductController {
 	
-	private final ProductService productService;
+	private final SpbDao spbDao;
 	private final ObjectMapper mapper;
 
 	@PostMapping("/addProduct")
@@ -30,36 +30,36 @@ public class ProductController {
 			System.out.println(mapper.writeValueAsString(product));
 		} catch (Exception ex) {
 			System.out.println(ex);}
-		return productService.saveProduct(product);
+		return spbDao.saveProduct(product);
 	}
 
 	@PostMapping("/addProducts")
 	public List<Product> addProducts(@RequestBody List<Product> products) {
-		return productService.saveAllProducts(products);
+		return spbDao.saveAllProducts(products);
 	}
 
 	@GetMapping("/fetchAll")
 	public List<Product> getAllProducts() {
-		return productService.getAllProducts();
+		return spbDao.getAllProducts();
 	}
 
 	@GetMapping("/fetchById/{id}")
 	public Product getById(@PathVariable int id) {
-		return productService.getProductById(id);
+		return spbDao.getProductById(id);
 	}
 
 	@GetMapping("/fetchByName/{name}")
 	public List<Product> getByName(@PathVariable @NonNull String name) {
-		return productService.getProductByName(name);
+		return spbDao.getProductByName(name);
 	}
 
 	@PutMapping("/updateProduct")
 	public Product modifyProduct(@RequestBody Product product) {
-		return productService.updateProduct(product);
+		return spbDao.updateProduct(product);
 	}
 
 	@DeleteMapping("/delete/{id}")
 	public String deleteId(@PathVariable int id) {
-		return productService.deleteProduct(id);
+		return spbDao.deleteProduct(id);
 	}
 }
